Fix Norm

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: mchaibi <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/13 13:25:04 by mchaibi           #+#    #+#             */
-/*   Updated: 2024/11/18 20:09:51 by mchaibi          ###   ########.fr       */
+/*   Updated: 2024/12/09 10:32:11 by mchaibi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,12 +19,11 @@
 int		ft_printf(const char *format, ...);
 int		ft_putunbr(unsigned int n);
 int		ft_handle_format(char c, va_list args);
-int		ft_puthex(unsigned long n, char x);
+int		ft_puthex(unsigned int n, char x);
 int		ft_putchar(char c);
 int		ft_putnbr(int n);
 int		ft_putstr(char *s);
 size_t	ft_strlen(const char *s);
-int     ft_putptr(unsigned long nb);
-
+int		ft_putptr(unsigned long nb);
 
 #endif
